Try to Fix Code Climate
