+++ app/Mail/Vaccine/VerifiedEmailNotification.php
@@ -66,6 +66,21 @@ class VerifiedEmailNotification extends Mailable
                 $this->subject = '[INFO] Permohonan Logistik Vaksin - Telah Realisasi';
                 $this->setMessageFinalized();
                 break;
+
+            case VaccineRequestStatusEnum::booked():
+                $this->subject = '[INFO] Permohonan Logistik Vaksin - Telah Dipacking';
+                $this->setMessageBooked();
+                break;
+
+            case VaccineRequestStatusEnum::do():
+                $this->subject = '[INFO] Permohonan Logistik Vaksin - Telah Dikemas';
+                $this->setMessageDo();
+                break;
+
+            case VaccineRequestStatusEnum::intransit():
+                $this->subject = '[INFO] Permohonan Logistik Vaksin - Paket Dalam Perjalanan';
+                $this->setMessageIntransit();
+                break;
         }
     }
 
@@ -140,16 +155,76 @@ class VerifiedEmailNotification extends Mailable
         $this->texts[] = '';
         $this->texts[] = '<b>Status Barang</b>';
 
-        $this->table = VaccineProductRequest::with('vaccineProduct')
-            ->where('vaccine_request_id', $this->vaccineRequest->id)
-            ->get();
+        $this->setTableValue();
 
-        foreach ($this->table as $key => $val) {
-            $this->table[$key]->finalized_status = $this->getStatusValue($val->finalized_status);
+        $this->notes[] = '<b>Lacak Permohonan</b>';
+        $this->notes[] = 'Lacak permohonan Anda melalui nomor Whatsapp Admin Logistik Vaksin Pikobar <a href="bit.ly/AdmLogVaksin">bit.ly/AdmLogVaksin</a>';
+        $this->notes[] = '';
+    }
 
-        }
-
+    public function setMessageBooked()
+    {
         $this->texts[] = '';
+        $this->texts[] = 'Terimakasih telah melakukan permohonan pada Aplikasi Vaksin Pikobar. Berikut merupakan ringkasan informasi permohonan Anda.';
+        $this->texts[] = '';
+        $this->texts[] = '<b>Status Permohonan</b>';
+        $this->texts[] = 'Melalui email ini, kami mengabarkan bahwa permohonan Logistik vaksin Anda dengan ID permohonan (' . $this->vaccineRequest->id . ') <b><i>telah dipacking.</i></b>';
+        $this->texts[] = '';
+        $this->texts[] = '<b>Tahapan Selanjutnya</b>';
+        $this->texts[] = 'Permohonan Logistik Vaksin Anda sedang dikemas untuk dimasukkan ke dalam kendaraan. Selanjutnya, akan dilakukan pengiriman barang sesuai tanggal rencana kirim.';
+        $this->texts[] = '';
+        $this->texts[] = '<b>Tanggal Rencana Kirim</b>';
+        $this->texts[] = 'Barang akan dikirim pada tanggal <b><i>' . date('d-m-Y', strtotime($this->vaccineRequest->delivery_plan_date)) . '</i></b>. Mohon bersiap pada tanggal tersebut.';
+        $this->texts[] = '';
+        $this->texts[] = '<b>Status Barang</b>';
+
+        $this->setTableValue();
+
+        $this->notes[] = '<b>Lacak Permohonan</b>';
+        $this->notes[] = 'Lacak permohonan Anda melalui nomor Whatsapp Admin Logistik Vaksin Pikobar <a href="bit.ly/AdmLogVaksin">bit.ly/AdmLogVaksin</a>';
+        $this->notes[] = '';
+    }
+
+    public function setMessageDo()
+    {
+        $this->texts[] = '';
+        $this->texts[] = 'Terimakasih telah melakukan permohonan pada Aplikasi Vaksin Pikobar. Berikut merupakan ringkasan informasi permohonan Anda.';
+        $this->texts[] = '';
+        $this->texts[] = '<b>Status Permohonan</b>';
+        $this->texts[] = 'Melalui email ini, kami mengabarkan bahwa permohonan Logistik vaksin Anda dengan ID permohonan (' . $this->vaccineRequest->id . ') <b><i>telah dikemas.</i></b>';
+        $this->texts[] = '';
+        $this->texts[] = '<b>Tahapan Selanjutnya</b>';
+        $this->texts[] = 'Permohonan Logistik Vaksin Anda sedang dimasukkan ke dalam kendaraan. Selanjutnya, akan dilakukan pengiriman barang sesuai tanggal rencana kirim.';
+        $this->texts[] = '';
+        $this->texts[] = '<b>Tanggal Rencana Kirim</b>';
+        $this->texts[] = 'Barang akan dikirim pada tanggal <b><i>' . date('d-m-Y', strtotime($this->vaccineRequest->delivery_plan_date)) . '</i></b>. Mohon bersiap pada tanggal tersebut.';
+        $this->texts[] = '';
+        $this->texts[] = '<b>Status Barang</b>';
+
+        $this->setTableValue();
+
+        $this->notes[] = '<b>Lacak Permohonan</b>';
+        $this->notes[] = 'Lacak permohonan Anda melalui nomor Whatsapp Admin Logistik Vaksin Pikobar <a href="bit.ly/AdmLogVaksin">bit.ly/AdmLogVaksin</a>';
+        $this->notes[] = '';
+    }
+
+    public function setMessageIntransit()
+    {
+        $this->texts[] = '';
+        $this->texts[] = 'Terimakasih telah melakukan permohonan pada Aplikasi Vaksin Pikobar. Berikut merupakan ringkasan informasi permohonan Anda.';
+        $this->texts[] = '';
+        $this->texts[] = '<b>Status Permohonan</b>';
+        $this->texts[] = 'Melalui email ini, kami mengabarkan bahwa permohonan Logistik vaksin Anda dengan ID permohonan (' . $this->vaccineRequest->id . ') sedang <b><i>dalam perjalanan.</i></b> menuju alamat tujuan.';
+        $this->texts[] = '';
+        $this->texts[] = '<b>Tahapan Selanjutnya</b>';
+        $this->texts[] = 'Selanjutnya, akan dilakukan proses serah terima barang. Mohon bersiap di lokasi kirim.';
+        $this->texts[] = '';
+        $this->texts[] = '<b>Tanggal Kirim</b>';
+        $this->texts[] = 'Barang dikirim pada tanggal <b><i>' . date('d-m-Y', strtotime($this->vaccineRequest->delivery_plan_date)) . '</i></b>. Mohon bersiap pada tanggal tersebut.';
+        $this->texts[] = '';
+        $this->texts[] = '<b>Status Barang</b>';
+
+        $this->setTableValue();
 
         $this->notes[] = '<b>Lacak Permohonan</b>';
         $this->notes[] = 'Lacak permohonan Anda melalui nomor Whatsapp Admin Logistik Vaksin Pikobar <a href="bit.ly/AdmLogVaksin">bit.ly/AdmLogVaksin</a>';
@@ -186,5 +261,14 @@ class VerifiedEmailNotification extends Mailable
                 break;
         }
         return $result;
+    }
+
+    public function setTableValue()
+    {
+        $this->table = VaccineProductRequest::with('vaccineProduct')->where('vaccine_request_id', $this->vaccineRequest->id)->get();
+        foreach ($this->table as $key => $val) {
+            $this->table[$key]->finalized_status = $this->getStatusValue($val->finalized_status);
+        }
+        $this->texts[] = '';
     }
 }
